Move query_translations.py to metrics/util
