+++ metrics/file_length.py
@@ -2,7 +2,7 @@ from byoqm.metric.metric import Metric
 from byoqm.metric.result import Result
 from byoqm.metric.violation import Violation
 from byoqm.source_repository.source_repository import SourceRepository
-from byoqm.source_repository.query_translations import translate_to
+from metrics.util.query_translations import translate_to
 
 
 class FileLength(Metric):
